renaming bunch of files, cleaning the README, and cleaning postprocessing a little bit
